+++ R/slope.R
@@ -45,6 +45,8 @@ slope = function(x, na_flag = -9999) {
   # Checks
   x = check_one_attribute(x)
   x = check_2d(x)
+  stopifnot(is.numeric(na_flag))
+  stopifnot(length(na_flag) == 1)
 
   # Make template
   template = x
@@ -108,3 +110,4 @@ slope = function(x, na_flag = -9999) {
   return(template)
 
 }
+